pyproject.toml
[project]
name = "webapp-clean-archi"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.8",
    "sqlalchemy>=2.0",
]

[dependency-groups]
dev = [
    "coverage>=7.7",
    "faker>=37.1",
    "mypy>=1.15",
    "pre-commit>=4.1",
    "pyright>=1.1",
    "pytest>=8.3",
    "ruff>=0.11",
]

[tool.ruff]
output-format = "concise"

[tool.ruff.lint]
select = [
    "E", "W",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
    "PL",  # Pylint
    "RUF",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"

[tool.mypy]
strict = true
disable_error_code = [
    "prop-decorator",
]

[[tool.mypy.overrides]]
module = ["tests.*", "app.api.routes.*"]
disable_error_code = [
    "no-untyped-def",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)30s - %(message)s"
